Completed the remaining models refactor by putting all the models into their own files
