+++ dna/models/dag_lstm_regression_model.py
@@ -1,0 +1,55 @@
+from .lstm_model import LSTMModel
+from .torch_modules.dag_lstm_mlp import DAGLSTMMLP
+from dna.data import group_json_objects
+
+
+class DAGLSTMRegressionModel(LSTMModel):
+
+    def __init__(
+            self, activation_name: str, hidden_state_size: int, lstm_n_layers: int, dropout: float,
+            output_n_hidden_layers: int, output_hidden_layer_size: int, use_batch_norm: bool, use_skip: bool = False,
+            reduction: str = 'mean', *, device: str = 'cuda:0', seed: int = 0
+    ):
+        super().__init__(
+            activation_name, hidden_state_size, lstm_n_layers, dropout, output_n_hidden_layers,
+            output_hidden_layer_size, use_batch_norm, use_skip=use_skip, device=device, seed=seed
+        )
+
+        self.reduction = reduction
+
+    def fit(
+            self, train_data, n_epochs, learning_rate, batch_size, drop_last, validation_ratio, patience, *, output_dir=None,
+            verbose=False
+    ):
+        # Get all the pipeline structure for each pipeline structure group before encoding the pipelines
+        self.pipeline_structures = {}
+        grouped_by_structure = group_json_objects(train_data, self.batch_group_key)
+        for (group, group_indices) in grouped_by_structure.items():
+            index = group_indices[0]
+            item = train_data[index]
+            pipeline = item[self.pipeline_key][self.steps_key]
+            group_structure = [primitive[self.prim_inputs_key] for primitive in pipeline]
+            self.pipeline_structures[group] = group_structure
+
+        super().fit(
+            train_data, n_epochs, learning_rate, batch_size, drop_last, validation_ratio, patience,
+            output_dir=output_dir, verbose=verbose
+        )
+
+    def _get_model(self, train_data):
+        return DAGLSTMMLP(
+            lstm_input_size=self.num_primitives,
+            lstm_hidden_state_size=self.hidden_state_size,
+            lstm_n_layers=self.lstm_n_layers,
+            dropout=self.dropout,
+            mlp_extra_input_size=len(train_data[0][self.features_key]),
+            mlp_hidden_layer_size=self.output_hidden_layer_size,
+            mlp_n_hidden_layers=self.output_n_hidden_layers,
+            output_size=1,
+            mlp_activation_name=self.activation_name,
+            mlp_use_batch_norm=self.use_batch_norm,
+            mlp_use_skip=self.use_skip,
+            reduction=self.reduction,
+            device=self.device,
+            seed=self._model_seed,
+        )
